utils: cmtspeech_ofono_test added
A test app that serves as a minimal example of an app using
libcmtspeechdata with oFono integration for call signaling. It
listens for oFono call signaling on DBus system bus. When a call
is established, it sets up the voice path with libcmtspeechdata,
and loops downlink audio as-is to uplink. Remote party will get its
uplink looped back.

Signed-off-by: Kai Vehmanen <kai.vehmanen@nokia.com>